backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
sentence-transformers
scikit-learn
python-multipart
